# Add a cache to LaunchedURLClassloader to improve startup performance