@wilkinsona 2022-03-02
This looks very promising. Thank you, @zjulbj. Have you done any comparisons of heap usage with and without the cache?