@mhalbritter 2022-03-07
Thanks for the heap comparisons! I'm going to bring this up on our next team call.