@mhalbritter 2022-03-10
Yes, that was meant with "indexing the classpath" :)