@zjulbj 2022-03-10
> Yes, that was meant with "indexing the classpath" :)

OK，thanks a lot. It seems this issue has been discussed long time ago  :)
http://mail.openjdk.java.net/pipermail/core-libs-dev/2015-August/035009.html
http://mail.openjdk.java.net/pipermail/core-libs-dev/2015-August/035010.html
http://mail.openjdk.java.net/pipermail/core-libs-dev/2015-August/035011.html
http://mail.openjdk.java.net/pipermail/core-libs-dev/2015-August/035012.html
http://mail.openjdk.java.net/pipermail/core-libs-dev/2015-August/035013.html
http://mail.openjdk.java.net/pipermail/core-libs-dev/2015-September/035192.html
http://mail.openjdk.java.net/pipermail/core-libs-dev/2015-September/035210.html
http://mail.openjdk.java.net/pipermail/core-libs-dev/2015-September/035290.html