@wilkinsona 2022-08-01
This will be 3.0.x at the earliest, but more likely later in the 3.x cycle. I have updated the milestone accordingly.