@wuyupengwoaini 2023-03-23
@wilkinsona @mhalbritter 
Is there any progress in "indexing the classpath"？I really want to know how this is going, because it seems to really help us shorten the startup time of the application. Because some of our big applications have thousands of dependencies