@wilkinsona 2024-01-31
Those interested in this PR, it would be really useful to know if you've tried running your app using class data sharing (CDS), particularly with Java 21. We've seen some very promising results with Spring Boot apps and would like to know if you have a similar experience. You can learn more in [this blog post from @sdeleuze](https://spring.io/blog/2023/12/04/cds-with-spring-framework-6-1) which includes some information about [a script](https://github.com/sdeleuze/spring-cds-demo/blob/main/unpack-executable-jar.sh) that unpacks a Spring Boot executable jar into a format that's as CDS-friendly as possible.