@mmoayyed 2024-01-31
I wrote about my experience with CAS [here](https://fawnoos.com/2024/01/21/cas70x-efficient-deployments/).

I did not try the script that does the unpacking, but in a nutshell, what I saw (with a WAR file) was:

```bash
jar -xf build/libs/cas.war
cd build/libs
java org.springframework.boot.loader.launch.JarLauncher
...
...
...
<Started CasWebApplication in 7.693 seconds (process running for 9.285)>
```

Then:

```bash
jar -xf build/libs/cas.war
cd build/libs
java -cp "WEB-INF/classes:WEB-INF/lib/*" org.apereo.cas.web.CasWebApplication
...
...
...
<Started CasWebApplication in 6.873 seconds (process running for 7.469)>
```


and then with CDS:

```bash
java -XX:+AutoCreateSharedArchive -XX:SharedArchiveFile=cas.jsa \
  org.springframework.boot.loader.launch.JarLauncher
...
...
...
<Started CasWebApplication in 6.766 seconds (process running for 7.582)>
```

Not that much difference, but I can see what I might have missed with the script.