@sdeleuze 2024-01-31
Please check the guidelines printed when using the `unpack-executable-jar.sh` script.

It is expected they you get no significant gain when using the `JarLauncher`.

```
To start the application from that directory:

      $ java -jar run-app.jar

      TIP: To improve startup performance, you can create a CDS Archive with a single training run:

      $ java -XX:ArchiveClassesAtExit=application.jsa -Dspring.context.exit=onRefresh -jar run-app.jar

      Then use the generated cache:

      $ java -XX:SharedArchiveFile=application.jsa -jar run-app.jar

      See https://docs.spring.io/spring-framework/reference/integration/class-data-sharing.html for more details.
```