@mmoayyed 2024-01-31
Thanks! I made small modifications to the script so it can run against a WAR file (replaced `BOOT-INF` with `WEB-INF`). With that in mind, I am not sure if this is still decent feedback and [a good test](https://github.com/apereo/cas-overlay-template), but here we go:

```bash
unpack-executable-jar.sh -d tmp build/libs/cas.war
```

Then:

```bash
$ java -jar run-app.jar
.
.
- <Started CasWebApplication in 6.637 seconds (process running for 8.178)>
```

Then:

```bash
java -XX:ArchiveClassesAtExit=application.jsa -Dspring.context.exit=onRefresh -jar tmp/run-app.jar
java -XX:SharedArchiveFile=application.jsa -jar tmp/run-app.jar
.
.
- <Started CasWebApplication in 6.284 seconds (process running for 6.693)>
```