@sdeleuze 2024-02-07
I had a deeper look, this CAS application seems to perform a lot of processing at startup. To put things in perspective, on my powerful MacBook Pro M2, the CAS application takes 5.285 s to start (after disabling key generation at startup) while Petclinic JDBC takes 1.448 s to start. CDS is improving the loading of the classes, it can't optimize custom application processing at startup.

If we analyze the data points more in details (unpacked variant is done with the [unpack-executable-jar.sh](https://github.com/sdeleuze/spring-cds-demo/blob/main/unpack-executable-jar.sh) script which replicate #38276 behavior):

JARexe: Started CasWebApplication in 7.854 seconds (process running for 8.709)
Unpacked: Started CasWebApplication in 4.778 seconds (process running for 5.285)
Unpacked / CDS:Started CasWebApplication in 4.36 seconds (process running for 4.712)

JARexe: Started PetClinicApplication in 1.586 seconds (process running for 1.842)
Unpacked: Started PetClinicApplication in 1.299 seconds (process running for 1.448)
Unpacked CDS: Started PetClinicApplication in 0.859 seconds (process running for 0.978)
CDS + AOT: Started PetClinicApplication in 0.678 seconds (process running for 0.81)
Project Leyden + AOT: even more

So CDS seems to optimize pretty consistently the loading of the classes on the 2 samples (0,573 s reduction versus 0,47 s). This half of a second on a powerful machine turns to be multiple seconds when deployed on a cheap Cloud server.

I think the layout used by the script and proposed by #38276 is immune even without CDS to the performance issue discussed in this PR, since it does not use `LaunchedURLClassloader`. So using both system classloader + CDS provide a significant performance boost.

AOT could be used to optimize even more the Spring Boot configuration processing of this application, but I was not able to make it work with CAS due to some double bean registration (feel free to raise a bug on https://github.com/spring-projects/spring-framework if you think there is something to fix on Spring Framework side).

Cache Data Store (Project Leyden premain) will be able to optimize with a wider scope than just class loading (optionally combined with AOT). It will for example include JVM AOT cache and maybe some classpath check caches. For the rest, that will be to the application itself to make careful use of the processing steps happening before startup.